hello-world/src/main/java/io/codebyexample/helloworld/core/usecase/GreetUseCaseImpl.java
package io.codebyexample.helloworld.core.usecase;

import io.codebyexample.helloworld.core.entity.Greeting;
import io.codebyexample.helloworld.dataprovider.id.IdProvider;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

/**
 * @author huypva
 */
@Component
public class GreetUseCaseImpl implements GreetUseCase {

  @Autowired
  IdProvider idProvider;

  @Override
  public Greeting greet(String name) {
    int randId = idProvider.genId();
    return new Greeting(randId, String.format("Hello %s!", name));
  }
}
